[IMP] Añadida primera aproximación al tablero

+++ app/src/main/java/com/example/protasks/models/ITask.kt
@@ -7,7 +7,7 @@ interface ITask {
 
     fun getPosition(): Int?
 
-    fun getId(): Int?
+    fun getId(): Long?
 
     fun setTitle(title:String)
